Rename data characterization workflow file.
